Moving TermFrequency to stats.
