+++ src/main/scala/nodes/stats/TermFrequency.scala
@@ -1,4 +1,4 @@
-package nodes.nlp
+package nodes.stats
 
 import pipelines.Transformer
