+++ src/test/scala/nodes/misc/TermFrequencySuite.scala
@@ -1,6 +1,6 @@
 package nodes.misc
 
-import nodes.nlp.TermFrequency
+import nodes.stats.TermFrequency
 import org.apache.spark.SparkContext
 import org.scalatest.FunSuite
 import pipelines.LocalSparkContext
